implemented firebase with fake message sending
