+++ server/package.json
@@ -13,6 +13,7 @@
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "faker": "^5.1.0",
     "firebase-admin": "^9.3.0",
     "google-auth-library": "^6.1.2",
     "mongodb": "^3.6.2",
